requirements.txt
# GitHub Tech Trends AI System - Dependencies

# Web Framework
fastapi==0.115.6
uvicorn==0.34.0

# HTTP Client
httpx==0.28.1

# Web Scraping
beautifulsoup4==4.12.3

# AI / ML
sentence-transformers==3.3.1
scikit-learn
hdbscan==0.8.40
numpy==2.2.2
pandas==2.2.3

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1

# Gemini AI
google-genai

# Phase 6: Optimization & Testing
fastapi-cache2[redis]==0.2.2
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-httpx==0.30.0
pytest-cov==5.0.0
